+++ apps/opentelemetry/src/otel_simple_processor.erl
@@ -91,7 +91,7 @@ init([Args]) ->
     Resource = otel_tracer_provider:resource(),
 
     {ok, idle, #data{exporter=undefined,
-                     exporter_config=maps:get(exporter, Args, undefined),
+                     exporter_config=maps:get(exporter, Args, none),
                      resource = Resource,
                      handed_off_table=undefined,
                      exporting_timeout_ms=ExportingTimeout}, [{next_event, internal, init_exporter}]}.
@@ -164,10 +164,10 @@ new_export_table(Name) ->
                     {write_concurrency, true},
                     duplicate_bag,
                     %% OpenTelemetry exporter protos group by the
-                    %% instrumentation_library. So using instrumentation_library
+                    %% instrumentation_scope. So using instrumentation_scope
                     %% as the key means we can easily lookup all spans for
-                    %% for each instrumentation_library and export together.
-                    {keypos, #span.instrumentation_library}]).
+                    %% for each instrumentation_scope and export together.
+                    {keypos, #span.instrumentation_scope}]).
 
 export_span(Span, #data{exporter=Exporter,
                         resource=Resource}) ->